fix CharGrid cell and text alignment issues
This fixes the uneven grid sizes, and text characters almost being drawn
in the entirely wrong cell.

I still don't know if '@' is a good example character for the
fixed-width font grid, but all of the other characters I could think of
had really strange metrics, so that's what we're going with for now.