replace CharGrid per-char atlas render with full grid texture rendering
Instead of drawing each character using GlyphCache, create our own image
buffer, render glyphs manually and update and draw a single texture for
the whole CharGrid.  This leads to a *drastic* reduction in CPU usage.

CharGrid rendering is now diff-based as well, so only changed cells need
to be re-rendered.  On top of that, if no changes are made to a CharGrid
between draw calls, no deep checking for changes is needed, and the
texture already present on the GPU is just redrawn.