create InputBuffer, let main loop become active/inactive as needed
Buffered inputs are needed to avoid inputs getting eaten when an input
event occurs but it's not the player's turn.

The main loop now has the ability to switch between actively looping and
emitting update events, and becoming inactive, waiting for inputs.  This
will very likely be needed for state-driven turn-based event flow later
on.