replace glutin_window with custom sdl2_window
pistoncore-glutin_window relies on an old version of glutin, which in
turn relies on an old version of winit, which panics on the latest
stable version of Rust when trying to generate a mouse movement event
due to using mem::uninitialized on a struct, a use case that is
apparently no longer supported by Rust.

The stock pistoncore-sdl2_window is not a drop-in replacement for
pistoncore-glutin_window, because it suppresses key repeat when a key is
held down.  The custom version I've patched into the Cargo.toml comments
this out, thus enabling the same key repeat that
pistoncore-glutin_window has.

Switching to pistoncore-sdl2_window also makes the font rendering much
crisper (perhaps pistoncore-glutin_window was doing anti-aliasing?), and
halves the disk size of the debug and release compiles.