calculate field of view with fixed-point iterative shadow casting
Each call to `field_of_view` requires allocating two Vecs, but otherwise
it should be pretty fast overall.

Visibility is generally considered center-to-center, but the
implementing map can choose to make certain tiles such as walls
asymmetrically visible for long walls to apear along corridors and
pressed against large walls.

Opaque tiles are considered diamond-shaped, which allows some corner
peeking and produces some nice shadows behind columns.

It took a _long_ time to settle on this approach and implementation,
especially when considering the topic of symmetry, where two tiles both
appear in each other's fields of view.  In particular, the idea of
symmetry in a field of view requires thinking about how lines of sight
will be approximated.

It is impossible to avoid "orphan" tiles (visible tiles that are
separated from other visible tiles) with certain patterns of opaque
tiles without also broadening visibility from center-to-center to
any-point-to-any-point, which can be done with some sort of precise
permissive field of view calculation.  However, this requires line of
sight to resolve linear separability of points using either some sort of
rotating calipers approach with convex hulls, or a perceptron/support
vector machine, neither of which have fast or precise algorithms.  Thus,
we accept "orphans" as a trade-off to make future line of sight
calculations simpler and faster.

Rust's borrow checker made this somewhat challenging to implement with
regards to performance, and I'm still not fully happy with it.  It
yelled at my page-flipping approach to managing sights for successive
rows because it didn't like that I was reading from one Vec and writing
to another as part of a 2-element array, so I was forced to decouple
them by duplicating some code and splitting the main scanning logic into
its own function.  It also forces a HashSet allocation for each fresh
field of view calculation (outside of `field_of_view`), because it needs
to borrow the `map` field of the `Game` struct, which means the callback
closure can't fill in the `fov` field of the same `Game` struct, even
though the two fields are disjoint, since the callback has to be `FnMut`
in order to do anything mutable.  This also stops access to the `map`
field in the callback, which would make `is_asymmetrically_visible`
unnecessary.  I think the whole source code needs some un-struct-ifying,
which would make some things a lot simpler and require less
dereferencing.