replace callback-based field of view with iterator approach
This eliminates the need for a callback closure when going through tiles
in the field of view, which crucially decouples the iteration body from
the map.  This means it's possible to e.g. check map properties during
the loop while adding returned coordinates to a data structure; this
entirely sidesteps the per-turn HashSet allocation the old
callback-based approach had to do.

In order to make this happen, I had to do the logical equivalent of
turning *four* nested loops inside-out, each of which had special
pre-loop logic that had to be done.  Only careful use of the Option type
managed to make this possible; I can't even imagine doing this without
it.