clean up FovIter::advance and FovIter::next, lower fov_shape check
The loop in FovIter::next was extremely ugly, so I had to do something.

The general field of view shape distance check has been lowered into the
core of FovIter::advance, so if it's circular it will now skip tiles
near the corners of the total potential square of the field of view.