simplify ruggle::run, remove App trait and AppContext
Instead of asking for an implementation of App with a solitary update
function, ruggle::run now just takes a closure and runs that repeatedly.

I also took the liberty of eliminating AppContext in favor of just
sending the InputBuffer and CharGrid directly to the update closure.
I did it to simplify the code, but it seems to improve CPU usage too.