movable player and creatures with Shipyard ECS
Back to following the Rust Roguelike Tutorial.  This commit covers:

 * creating Position and Renderable components
 * spawning a player and some creatures with those components
 * rendering the Renderables at the correct Positions
 * moving the creatures to the left with a tag component and system
 * moving the player by manually altering components in the world

This is combined with the pre-existing work with the map, field of view
calculation and vision centered on the player.

I'm trying out Shipyard as the ECS instead of Specs; apparently it's
faster for a bunch of key use cases, and has less boilerplate code.