move lib modules into their own subdirectory
It was way harder to figure out how to do this than necessary.
Apparently just changing "lib.rs" to "lib/mod.rs" is not recognized out
of the box, so I had to manually put a "[lib]" section in `Cargo.toml`
with the path I wanted; figured I may as well use "lib/lib.rs" if it's
going to make me specify a path manually anyway.