integrate fallback logic into find_path, unify bounds logic
Fallback when failing to find an exact path is now done by sending a
flag to find_path.  This performs the A* search starting from the start
position, checking which of the reachable positions is closest to the
destination, and returning a path to the that position instead.

Sub-bounds no longer takes points explicitly; instead it takes a padding
value and assumes the desired bounds will be a rectangle containing the
start and destination positions.

These two things combined makes using find_path a whole lot simpler.