let monsters attack the player, handle player death
Monsters now attack the player via the same code path as the player
attacking monsters.  If the player dies, a message is logged, normal
game flow is stopped, and nothing will happen until space is pressed,
which quits the game.

Whether or not the player is alive is now tracked, and
delete_dead_entities now skips player entities.

To allow the game to quit on space, the callback sent to ruggle::run now
returns two bools, one to keep running and the other for continuous
updates (as before).

RunSettings::start_inactive has been removed; it turns out that the same
preference can be inferred from the first run of the callback sent to
ruggle::run.