implement mode stack to control high-level game logic
I really didn't like the Epic Mega State of Doom that the Rust Roguelike
Tutorial was devolving into when it started showing an inventory menu.

The mode stack is my solution to this.  With a mode stack, I can keep
the main dungeon mode fully separate from the upcoming inventory mode.
In general, the mode stack keeps the update and drawing logic for each
mode close to each other, while separating them from the updating and
drawing of other modes.  The mode stack only updates the top-most mode,
so I don't need to carefully avoid updating other parts of the game by
mistake.  The mode stack also draws its modes layered from the
bottom-up, and if a mode isn't on the stack it won't be drawn, so no
weird conditions to do things like avoiding rendering entities while a
title screen is showing, for example.

See `src/modes/mod.rs` for more details.