allow player to pick up, drop and view inventory items
This introduces three new modes: PickUpMenuMode, InventoryMode, and
InventoryActionMode.

PickUpMenuMode shows a menu of the items at the player's position and
allows them to choose one to pick up off the ground and add to their
inventory.

InventoryMode shows a list of the items in the player's inventory.  The
"Sort all items" option does nothing right now, and the empty spaces
will eventually show the character status and equipment.

Picking an item in the InventoryMode presents an InventoryActionMode
with the item and possible actions to perform with it; right now the
only option is "Drop".  Choosing "Drop" will pass the choice back to
InventoryMode, which in turn passes it back to DungeonMode, which drops
the item onto the map from the player's inventory.