add CharGrid::draw_bar, use it for HP, pick up, inv + inv position fix
CharGrid::draw_bar can draw a partially filled bar, so it's now used to
draw the HP bar at the bottom of the screen, as well as the new scroll
bars in the inventory and pick up item lists.

Also, the inventory item list area was one space too short due to the
top equip area taking an extra space than needed, so give it back to the
item list.