allow player to auto run by pressing Shift+direction
Auto run comes in two flavors: corridors and straight lines.  Corridor
auto run follows a corridor, including around turns of up to and
including 90 degrees left or right.  Straight line auto run may follow a
straight wall until it changes, or proceed into the open until any wall
is encountered.

Auto run is prevented if any enemies are present in the player's field
of view.  Auto run will stop if the player presses a key or becomes
adjacent to something interesting such as an item on the floor.

In the future, the auto run system could be extended to handle things
like resting in place to heal and traveling to a specific location.