add magic missile scroll with ranged targeting UI
The player can now find magic missile scrolls.  Using a scroll brings up
a targeting mode that allows the player to damage monsters from afar.

The logic for items that cause damage is very simple, resembling the
healing effect of potions, but in reverse.

The target mode interface is the majority of this commit, and is an
excellent example of the use of a mode stack to keep non-trivial logic
isolated from the rest of the game while still being able to interact
with it.  Choosing to apply an item with a Ranged component causes the
InventoryActionMode to push the new TargetMode onto the stack.
TargetMode draws over the whole screen, so a `draw_behind` function is
introduced to Mode to tell the ModeStack which modes in the stack should
be drawn behind others.  If the player chooses a target, the target
coordinates are returned, which gets bundled with the item to use by
InventoryActionMode.  If the player targets themselves, a
YesNoDialogMode is pushed to ask the player to confirm their action.
Thanks to the mode stack construct, cancelling out of any of these modes
returns the player to the underlying modes, with full memory of the
state of menus and the like.

This would have been a huge mess in the big state machine approach used
by the Rust Roguelike Tutorial, and it was this exact feature that
motivated me to come up with the mode stack concept to begin with, so
I'm really glad that it worked out as well as it did.