add fireball scroll with area of effect
The targeting interface shows the area of effect of fireball scrolls and
will warn the player when they are included in that area when choosing a
position.

`spawn_fireball_scroll` bumps into a limit in the number of storage
arguments that Shipyard can handle at once; ten max.  Fortunately,
Shipyard permits nested tuples to get around this, which is what is used
here, but it looks a bit strange.