add sleep scroll that inflicts asleep status effect
Sleep scrolls have a 1-tile area of effect, and inflict 36 'ticks' of
sleep, deducted as follows: -1 each turn, an extra -1 if the opposing
player or monster is within the sleeper's field of view, and an extra
-10 if the sleeper took damage in the past turn.  In practice, a sleep
scroll can be used to get 4 free attack turns against a monster, or gain
around 17 free turns in a monster's field of view without it moving.

If the player is asleep, pressing any key other than escape (which
brings up the save and exit dialog) will pass a turn.