migrate off of piston and onto sdl2
There was nothing unique or easier about what Piston was being used for
versus any other game library or framework, and SDL 2 was already being
used as the Piston backend, so now SDL 2 is just being used directly
instead of through an intermediary layer.  The biggest benefit to this
is shedding a whole lot of dependencies that Piston was pulling in.

There are four major parts to the migration: setting up the window, the
main loop, handling input and handling output.  The window setup and
input handling are fairly straightforward.  However, the Rust SDL 2
crate requires textures to be created from a texture creator, which it
must not outlive.  The CharGrid holds a reference to this texture, which
means it must be created *after* the texture creator, and thus the
window.  However, the CharGrid was being created early to load the font
and determine grid dimensions in order to set the initial window size.
This cyclic dependency had to be broken: the font image is loaded in
advance, and grid dimensions are passed into a CharGrid-associated
helper function to determine the initial window size, and then the
CharGrid is created, consuming the font image during construction.

The main loop has to do a whole bunch of timing to prevent updating too
quickly, updates falling behind or spinning the overall loop too
quickly.  One side effect of this is that the game is *much* slower.  It
turns out that I was misusing Piston this entire time, and only limited
the redraw rate by setting FPS; there is a separate "updates per second"
setting that I never touched because I didn't understand it until now,
but apparently the default UPS is 120 updates per second, so the game
has just been updating *way* faster tha intended up until now, and was
only rate-limiting its redraws.  The game logic itself needs to be
changed in order to improve its speed.