cargo update, fix bitvec -> funty failure to compile
bitvec 0.19 fails to compile with funty 1.2 due to "multiple applicable
items in scope" for `BITS`.  The root cause is bitvec's failure to lock
the minor version of its dependency on funty; in the mean time, the
funty version will be locked here to a version prior to 1.2 as a
workaround.