drop image crate in favor of sdl image, render with sdl surfaces
Font image loading and the back buffer for the drawn CharGrid texture
are now SDL 2 surfaces, for a pretty substantial rendering performance
boost.

As a bonus, dropping image drops the number of crates needed to compile
the game from 90-ish to about 60; when Piston was being used it was
nearly 200, so the overall fresh compile time and final binary size have
dropped by a lot over the last few days.