enable player to descend to deeper levels via downstairs
The game now places downstairs in one of the rooms of the level.  The
player can take the stairs down to the next level by pressing '>' or
enter and confirming the prompt that pops up.

To detect the greater-than key, the state of the shift key needs to be
known, so unfortunately the nice From trait can no longer be used to
easily convert from SDL keycode to the game's custom GameKeys.

Going down a level works by first despawning all entities that aren't
the player or their possessions; this bumps against the first limitation
I found in Shipyard while implementing this, which is that there is no
way to iterate over all entity IDs; I wasted a lot of time trying to
find a way before I reached that conclusion.  Looking at Shipyard's
source code, it wouldn't even be that hard to implement, it's just
straight up absent for whatever reason.

The next step in going down a level involves clearing the map, where I
hit the other Shipyard limitation: once a unique (or "resource" in other
ECS libraries) has been inserted into the world, you cannot replace it.
I would normally have cleared the map by creating a new map and
replacing the existing one, but Shipyard doesn't allow that; it doesn't
even allow removing a unique once added.  This can be worked around with
a proxy type that holds the type you actually want with a replace
function, but it's pretty clunky for no good reason.

Anyway, after the despawning and map clearing, the map is generated
anew, spawns are filled in and the player (who still exists in the
world) is appropriately placed with their field-of-view dirty flag set
to properly show the player's new surroundings.