change Size width and height from i32 to u32
This is closer to the convention used by the rust-sdl2 crate.

Generally, any u32 that stores dimensions that will eventually be
indexed with i32 will assert for >= 0 and <= i32::MAX at the time they
would be assigned or changed.  This makes it safe to convert those u32s
back into i32s when needed without having to use TryInto and checking
all over the place.