manually set glyph size and char mapping with FontInfo
FontInfo takes the place of the former free-standing PathBuf to the font
image, allowing the pixel width and height of glyphs and the mapping of
characters to glyph positions on the image to be manually specified.

To help with loading a font with the standard 16-by-16 layout of IBM
Code Page 437 glyphs, FontInfo::map_code_page_437 can create a
ready-to-use character map in a single call.

Loading a font image performs some basic validation that the glyph
positions are within the bounds of the font image.