set window size in pixels, move minimum size enforcement out of CharGrid
This was leading to silly things like sending in minimum grid dimensions
to the CharGrid system and then immediately querying it for actual pixel
dimensions when the initiator was already working in pixel space.

Handling minimum size is now the responsibility of whoever is handling
the CharGrid; for now this is ruggle::run, which now takes pixel sizes
for the initial and minimum window sizes instead of grid dimensions.