manage multiple CharGrids in CharGridLayers, create and resize in modes
Instead of having a single CharGrid that everything draws onto and gets
displayed, CharGrids now live in CharGridLayers.  A vec of
CharGridLayers is given to ModeStack, which ensures that there is a
CharGridLayer for each mode in the stack.  Each mode can then create as
many CharGrids in that CharGridLayer as it needs, and will then become
responsible for resizing and placing those CharGrids on screen by
manipulating their corresponding CharGridViews.

The dungeon and target views now use a CharGrid each for the map and the
user interface.  The inventory screen now splits the equipment, status
and inventory list into three separate CharGrids, drastically
simplifying the drawing code by eliminating lots of incidental offsets
that previously had to be included for correct placement of drawn
characters.