handle app quit event
When the game receives a quit event, e.g. window close or Alt-F4, it now
shows a dialog asking to quit, instead of simply ignoring it.

This works by introducing InputEvent::AppQuit, an input event that can
be handled anywhere input is handled in general.  Modes on top of
DungeonMode will return their own individual AppQuit mode results if
they detect this input or any mode on top of them does, all the way back
down to DungeonMode.  DungeonMode itself detects AppQuit mode results
and pushes the new AppQuitDialog mode to ask the player if they really
want to quit.

As a special case, InputEvent::AppQuit quits out immediately when the
player is dead and the game asks for a key to continue.