remove TileImage, move tiles from image to 1-tile-wide surface instead
Instead of keeping a tile image in memory in roughly the same way as on
disk, take the tiles in the image and put them on a specialized
one-tile-wide surface.  This puts all the rows of pixels for a single
tile close in surface memory.  Doing this just brought the game back to
smooth 60 FPS in unzoomed debug mode again.

As part of this, the input font map for a Tileset now gets internally
converted to address y positions in the surface instead of indexing the
original image by tile.