move work out of TileGrid::render loop, page flip RawTileGrid cells
This moves a bunch of calculations that didn't need to happen in the
cell-checking loop out of that loop.  Anything that remained was moved
in after the cell comparison to make skipped cell logic as fast as
possible.  This doesn't have a noticable impact on performance, but it
makes the code cleaner as well, so it may as well be done.

Instead of writing the front RawTileGrid cells to the back RawTileGrid,
just swap the Vecs for the cells instead, since the front grid gets
completely overwritten every time drawing happens anyway.