split main loop from loop init, use rust-sdl2 unsafe_textures feature
ruggle::run now delegates its work to the new RunLoop struct whose
purpose is to split initialization (RunLoop::new) and the main loop body
(RunLoop::main_loop) into separate functions.  Unfortunately, this
pushes the fullscreen 8-by-8 Terminal font rendering below 60 FPS again,
but this is needed for an upcoming change.

To allow RunLoop to hold everything, the `unsafe_textures` feature of
rust-sdl2 must be enabled, otherwise Textures and TextureCreator would
have to live in the same struct, and since Rust doesn't track lifetimes
at the individual field level, there is no way to express to the borrow
checker that Textures never out-live the TextureCreator in the same
struct.  With `unsafe_textures`, TextureCreator no longer exists, so
Textures no longer have lifetimes, but they also have no Drop
implementation to clean themselves up.  To perform automatic clean-up, a
new GridTexture smart pointer holds Textures with a Drop implementation
that uses an unsafe SDL_DestroyTexture call on the raw SDL_Texture
pointer; this is safe as long as GridTextures are guaranteed to be
dropped while the canvas from which they are created is still alive,
which is ensured in this case by putting the TileGridLayers above the
Canvas in the RunLoop struct (Rust guarantees that struct fields are
dropped in declaration order).