increase default map zoom and window size, reduce FPS to 30 again
The native default window size has been increased from 800x600 to
1000x600, to match closer to the browser version.

The map zoom is now 2x zoomed by default.  This drastically reduces the
number of tiles that have to be rendered, which takes a lot of pressure
off of the CPU.

The FPS has been dropped to 30, also to reduce CPU pressure.