profile and optimize render::draw_map
render::draw_map is by far the most time-consuming function when moving
around the map, according to Firefox's profiler.  In response to this,
optimizations have been made to it (listed in order of perceived
impact):

 * Skipping unseen tiles earlier in Map::iter_bounds returning
   (GameSym, Color) instead of out Option<(GameSym, Color)>.  Impact is
   reduced when many tiles have been previously seen and thus have to be
   visited anyway.
 * Inlining functions involved in getting a map tile at a position.
 * Inlining functions involved in checking if a tile is within a field
   of view.

The performance improvement is quite noticeable, and although release
WebAssembly is still quite a bit slower than the native debug build,
it's playable with minimal lag even at 1x zoom with the 8-by-8 Terminal
font set as the map tileset.

The WebAssembly+Emscripten build is now set to emit profiling-friendly
output, but the fully-optimized flags are left commented out to easily
use them instead.