replace render::draw_map with chunked map grid drawing for performance
Profiling the WebAssembly+Emscripten build, over 35% of frame time is
spent drawing, i.e. iterating potential map spaces and transferring them
onto a TileGrid.  This commit replaces render::draw_map with an approach
that drops this to about 5% of frame time using:

 1. tiled dirty rectangles to only redraw map chunks when they get
    scrolled into view or by manual request, and
 2. wrapped offsetting to keep map chunks at the same place in the
    TileGrid whenever they should be visible while maintaining the
    illusion that the map view is centered about the player.

This is all facilitated by ChunkedMapGrid: a new struct that is stored
as part of a Mode and has methods that track map chunks, allow marking
chunks as dirty and selectively draws chunks as needed.  Currently, a
map chunk is an 8-by-8 tile square of map tiles; the logical map tile
storage has been padded and rearranged to make this cache-friendly.

A couple of earlier performance changes were undone to support this new
approach, which nevertheless still comes out ahead:

 1. Map::iter_bounds returns an Option again, so blank spaces can be
    manually filled.  Using TileGrid::clear prior to drawing just a
    small map chunk would undo our preserved drawing of other chunks.
 2. Page-flipping in TileGrid::render is replaced with a
    slice::copy_from_slice of front RawTileGrid cells to the back cells,
    since we no longer guarantee that all cells are drawn to every
    single time TileGrid::display is called.

The end result still doesn't quite reach 30 FPS in the
WebAssembly+Emscripten build, but it gets close enough that the game is
playable in a web browser even when fullscreen with the smallest font
used to render the map.