replace random corridors with corridors connecting closest rooms
This effectively turns the connections of rooms with corridors into a
minimum spanning tree using Prim's algorithm.  This removes the
excessive criss-crossing of corridors over other corridors and through
rooms which was otherwise really ugly.  On the other hand, it all but
guarantees that most rooms have only one path through, which takes away
the possibility of detouring tough rooms, so I'm unsure if I'll stick
with this in its pure minimum spanning tree form.