add experience gain and levels
Sources of damage now add a HurtBy component to the damaged entity.  If
an entity dies, whoever hurt it last is granted experience points,
stored in their Experience component.  If experience exceeds a certain
amount, the entity will gain a level, and if they have a CombatStats
component, it will increase their maximum hit points.