replace global RNG with seed + hashing with short-term RNGs
Random numbers produced for use with the game now come from a single
game seed.  Whenever a part of the game needs random numbers, this game
seed is hashed along with some other data to serve as the seed of an RNG
created on demand.  This allows the same game seed to e.g. generate the
same level layouts at the same depths, even if random numbers are
generated elsewhere in the code.