implement 10% miss chance for attacks, vary damage +-20%
Every melee attack now has a 10% chance to miss.

The damage inflicted by a melee attack can now be up to 20% above or
below its usual value.  Fractional damage is rounded probabilistically,
e.g. 3.1 damage is rounded to 4 damage 10% of the time and rounded to 3
damage 90% of the time.

The RNG for calculating all of this is seeded with the game seed, turn
count and attacker and defender coordinates; the last two require minor
changes to monster::do_turn_for_one_monster to prevent holding a mutable
borrow on Coords when the attack is performed.