scale attack, defense and hp by level
Attack, defense and hit points for the player and monsters are now
scaled by experience level, as well as attack and defense bonuses for
weapons and armor respectively.  For now, monsters, weapons and armor
are all calculated based on being level 1.  Eventually, monsters will be
spawned with increasingly higher levels corresponding to improved combat
stats, while new weapons and armor will be spawned with higher bonuses.

Scaling of attack and defense by level is new to this commit, so gaining
levels makes the player too powerful right now.