adjust player/equipment attack/defense, spawn knife+shield in first room
The first room of of the first dungeon level now has a guaranteed knife
and wooden shield.  The player's stats have been adjusted downwards and
equipment stats changed such that the starting player with a knife and
wooden shield has attack and defense comparable to being unarmed
previously.

In general, player stats and equipment bonuses are split 60-40 in favor
of player stats.  The exact split, as well as the level and frequency of
equipment upgrades in the future, still needs to be decided.