choose monster spawns based on difficulty
Monsters are now chosen based on difficulty, replacing the previous
random monster selection scheme.

The exact level of a spawned monster is split three ways:

 * 20%: depth-appropriate monster
 * 40%: above, but one to three levels lower
 * 40%: a randomly-chosen monster even weaker than the above

This is kind of tricky to see in action since there's only two monsters
right now, but otherwise it works, e.g. a spawned goblin will always be
sort of weak.