increase nutrition by 50%, undouble regen rate, remove full regen bonus
This increases the nutrition of rations and of the player's stomach by
50%, loosening the hunger clock by a lot.  Regeneration has been
unbuffed, and regeneration from having a full stomach has been nerfed to
match the normal regeneration rate instead of being doubled.

These changes allow players to back out of dangerous encounters to
explore other parts of the level, use items or even rest in place more
often, instead of mindlessly charging forwards to appease the food
clock.