implement word wrapping and use it for messages in the sidebar
This introduces ruggle::word_wrap, which returns an iterator that breaks
a long string into word-wrapped lines of at most a given number of
characters.  It should be as fast as it can be, considering that it only
scans forwards, while also breaking words longer than the line width,
respecting the variable-width UTF-8 encoding, leading and trailing
whitespace, and hyphens.

Word wrapping is applied to messages and prompts that appear in the
message area.

The maximum number of messages stored and possibly displayed has been
increased from 4 to 100.