support shortcut keys in inventory and inventory action modes
Pressing [e]quip (or [w]ield/[w]ear), [a]pply and [d]rop while in the
inventory mode will now bring up the inventory action mode with the
action pre-selected if the item supports that action.

Pressing those keys in the inventory action mode will confirm the action
if it is already selected, or otherwise switch the selection to it.