factor out Map::describe_pos from TargetMode
Factoring this description-building logic out allows it to be
generalized to provide sensible descriptions of map positions outside of
the player's field of view, which will be needed soon.

The new position description may produce slightly longer descriptions
than the old code, so the prompt in the message area is given an extra
line to ensure that it fits properly.