fix handling of negative values in chunked map rendering calculations
Bad Things were happening when chunks above or to the left of chunk
(0,0) of the map needed handling, all due to failing to account for
standard integer division rounding towards zero (the wrong thing) versus
towards negative infinity (the right thing).

Integer division is replaced with i32::div_euclid when calculating which
map chunk is supposed to be considered to be at the top-left of the
screen, as well as dirty rectangle bounds.

Calculation of how map chunks correspond to screen chunks needs special
attention, involving finding the lowest chunk multiple that is higher
than the absolute input value, adding it to guarantee a positive value,
and only then following the standard modulo arithmetic approach.

These fixes combined remove weird rendering artifacts when anything is
drawn at negative coordinates.