add view map mode to let the player look around by moving the camera
The player can now press 'v' to enter the view map mode, which is like
target mode without the targeting.  In this mode, the camera can be
moved around, a cursor is shown and a description of the tile under the
cursor is given in the message prompt area.

The movement keys move the cursor, and shift can be held to move faster.
Pressing Home recenters the camera on the player.