unify handling of waited and polled events in main game loop
This removes the redundant handle_event function to separately handle an
event that is waited on and subsequent polled events.  Taking the waited
event as an Option<sdl2::event::Event> and turning it into a single-item
iterator that chains into the polled event iterator allows all event
handling to be handled inline instead of being forced to use a helper
function.