show a game over screen when the player dies
Dying will now take the player to a separate game over screen, showing
their stats, item count and equipment, as well as damage dealt, damage
taken and kill count tallied over the course of the game.

Tallying has to be hard-coded in the various places where it takes
place, since the version of Shipyard used by the game does not support
custom arguments for systems.

HurtBy clearing is reduced to only happen when the player is alive at
the end of the turn, so whatever kills the player can be credited on the
game over screen.

Finally, there's a minor change to terminate dead entity handling
earlier to ensure that the prompt message after the player dies is
always the last message shown in the message log.