update Shipyard from 0.4.1 to 0.5.0
Shipyard 0.5.0 adds World::remove_unique, which is the motivation of
this entire version update.  Unfortunately, it fails to build for the
wasm32-unknown-emscripten target, so this commit will have to be
reverted; this commit therefore only exists for the sake of posterity.
What follows are a broad description of the changes needed to port from
Shipyard 0.4.1 to Shipyard 0.5.0, and a more detailed explanation as to
why this commit needs to be reverted.

This game mostly uses Shipyard as a entity-component store, with very
light usage of World::run and no usage of systems or workloads.  With
that in mind, porting from Shipyard 0.4.1 to Shipyard 0.5.0 involved the
following changes:

 * Shiperator::with_id is now IntoWithId::with_id.

 * World::borrow now returns a Result, which is resolved with unwrap.

 * World::run now returns a Result, which can be unwrapped to return
   whatever the function otherwise normally returns.

 * World::get now returns a Result, which is resolved with unwrap.

 * World::try_get no longer exists, and is replaced with World::get.

 * Entities::add_component now takes the entity ID as the first argument
   instead of the last.

 * Remove::remove no longer needs turbofish syntax and can now be just
   dot called off of a mutable view or tuple of mutable views.

 * AllStorages::delete is now AllStorages::delete_entity.

The reason this commit must be reverted is because it fails to build for
the wasm32-unknown-emscripten target.  Specifically, attempting such a
build raises the following error (with -s LLD_REPORT_UNDEFINED):

    wasm-ld: error: (...): undefined symbol: _emscripten_get_now

The error without the linker flag suggests "__emscripten_get_now may
need to be added to EXPORTED_FUNCTIONS if it arrives from a system
library", but doing so just complains about "wasm-ld: error: symbol
exported via --export not found: _emscripten_get_now", and trying close
alternatives like "_emscripten_get_now" and "emscripten_get_now" just
raise the first error.  The problem seems to be the new instant crate
(v0.1.10), pulled in via an update to parking_lot (v0.10.2 -> v0.11.1)
which is a required dependency of shipyard v0.5.0; the instant crate
must need emscripten_get_now via some call-chain, but hours of search
engine trawling, code-diving and experimenting with EXPORTED_FUNCTIONS
proved fruitless.  Apparently, parking_lot has a feature flag to not
require Instant::now, but toggling it requires nightly Rust.  The most
ironic thing about the instant crate being the show-stopper here is that
the whole point of the crate is to act as a compatibility layer to
provide functionality similar to the Rust standard library
std::time::Instant for WebAssembly, which doesn't support it directly.

To sum up the above: shipyard v0.5.0 unconditionally requires
parking_lot v0.11, which requires instant, which fails to find
emscripten_get_now at link time, most likely due to being dropped by the
optimizer, with no way to fix it; this blocks the
wasm32-unknown-emscripten target and therefore the WASM port of this
game, so this commit needs to be reverted.