add save and load support to the game
The game will now save to a save file when choosing "Save and exit" in
the options menu or when confirming the app quit dialog.  When a save
file is detected on the title screen, a "Load Game" option will appear
that will load the game from that save file and allow the player to
resume playing.  If "New Game" is chosen and a save file is present, a
dialog asking to delete the save file will appear instead.

The save data stored in "savegame.txt" is a structured text file where
each line represents some data to be loaded back into the world.  Each
line describes either a unique or an entity component.  A unique line
consists of the type name followed by its serialized data, while a
component line consists of the entity ID, a tab (used to distinguish
component lines from unique lines), the component type name and
serialized component data.  The process of loading involves parsing each
line; uniques are loaded into temporary variables, while components are
loaded and added to new temporary entities that are created if their ID
is newly seen.  Validation is performed, followed by a second pass to
translate saved IDs into loaded IDs for uniques and components that hold
entity IDs.  Finally, old entities are despawned and the game is put in
the main dungeon gameplay mode.  Saving is very simple in comparison:
all game world data worth saving is written out, one unique or
entity-component pair per line.

Some game data structures needed adapting for the saving and loading
processes.  Shipyard 0.4 lacks the ability to reassign or remove uniques
once they've been added, so several game structs stored as uniques now
have an associated `replace` function that just field-replaces  the
existing unique with a new instance that it consumes.  A `capacity`
field needed to be added to the `Messages` struct, since the up to this
point had been using the implicit capacity of the `VecDeque`.  The `id`
field of `Difficulty` has been made public to facilitate entity ID
fixing when loading the game.

Serialization is handled via the serde and ron crates, with RON (Rusty
Object Notation) being used as the serialization format.  The biggest
headaches/time-wasters in implementing this come from deserialization
failures when dealing with the two external data structures used by the
game, namely Shipyard 0.4's EntityId and bitvec's BitVec.

EntityId uses custom serialization and deserialization code, but it
serializes as a tuple struct while deserializing as a plain tuple.  This
discrepancy passes undetected in its own serde_json test, but fails with
ron when outputting structs with type names enabled, the feature of the
format I was originally planning to use to distinguish different unique
and component types at load time.  The current manual output and parsing
is a compromise, and this mandates implicit struct names instead that I
don't like for parsing ambiguity reasons, even though this doesn't
affect the game in its current state.

Meanwhile, deserialization of BitVec fails for reasons I didn't bother
investigating.  Instead, BitVecs are converted to run-length encoded
sequences of 1s and 0s using a custom adapter whose creation required
looking up a basic example on how to handle a struct field specially on
Stack Overflow; serde's documentation was surprisingly unhelpful for
this, even with its many stock examples.  The same run-length encoder is
used for map tiles, which reduces the save file size to a third of what
it would be otherwise.